% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugs.R
\name{debug_with_message}
\alias{debug_with_message}
\alias{debug_once_with_message}
\alias{undebug_with_message}
\title{Debug with a message}
\usage{
debug_with_message(fun, text = "", condition = NULL, signature = NULL)

debug_once_with_message(fun, text = "", condition = NULL,
  signature = NULL)

undebug_with_message(fun, signature = NULL)
}
\value{
a debuged function
}
\description{
Debug with a message
}
\examples{
\dontrun{
a <- function() "plop"
debug_with_message(a)
undebug_with_message(a)
}
}
