% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_back.R
\name{trace_with_message}
\alias{trace_with_message}
\alias{untrace_with_message}
\title{Trace a function}
\usage{
trace_with_message(what, tracer, exit, at, print, signature,
  where = topenv(parent.frame()), edit = FALSE)

untrace_with_message(what, tracer, exit, at, print, signature,
  where = topenv(parent.frame()), edit = FALSE)
}
\description{
Trace a function
}
