% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_back.R
\name{trace_back}
\alias{trace_back}
\title{A reverse traceback}
\usage{
trace_back()
}
\value{
a reveresed traceback
}
\description{
A convenient traceback that print the call stack in the order that it happens
  (unlike default traceback).
}
